+++ example/web-app/js/ui-event.js
@@ -1,0 +1,7 @@
+const eventset = require('eventset');
+
+var UIEvent = eventset.Topic('app-ui-event');
+
+UIEvent.addEvent('toggle-list');
+
+module.exports = UIEvent;