+++ changelog.md
@@ -3,9 +3,9 @@ All notable changes to this project will be documented in this file
 
 ## v1.7.1
 ### Fix
-- Topic.dispatch() : the event can now be dispatched without message,
-in this case the listeners argument will not contains message property
-ex : Topic.dispatch('show-list') , the second argument can be omited
+- **Topic.dispatch()** : the event can now be dispatched without message,
+in this case the listeners argument will not contains message property.  
+example : Topic.dispatch('show-list') // the second argument can be omited.
 
 ## v1.7.0
 ### Added